create spatial_neighs script, move _add_complexes
