+++ liana/method/sp/__init__.py
@@ -1,4 +1,4 @@
-from ._spatial_pipe import spatial_neighbors
+from ._spatial_neighbors import spatial_neighbors
 from ._lr_bivar import lr_bivar
 from ._bivar import bivar
 from ._bivariate_funs import show_functions
