+++ liana/tests/test_spatial_utils.py
@@ -3,7 +3,7 @@ from  scipy.sparse import csr_matrix
 
 from liana.method.sp._spatial_pipe import _global_zscore_pvals, _global_permutation_pvals, _local_permutation_pvals, _local_zscore_pvals
 from liana.method.sp._bivariate_funs import _local_morans
-from liana.method.sp import spatial_neighbors
+from liana.method.sp._spatial_neighbors import spatial_neighbors
 
 from liana.testing._sample_anndata import generate_toy_spatial
